# client struct: lazy init components and optimize struct layout