@zuiderkwast 2024-12-09
> * 45% smaller base client structure - Basic clients now use 528 bytes (down from 896).

Amazing!